Various bug fixes

+++ src/include/compiler/tokens.h
@@ -5,7 +5,7 @@
 #include "../runtime/value.h"
 
 #define IS_VALUE_TOK(TOK) (TOK.type < 13)
-#define IS_STATMENT_TOK(TOK) (TOK.type - 10 < 10)
+#define IS_STATMENT_TOK(TOK) ((unsigned int)(TOK.type - 10) < 10)
 
 struct token {
 	enum token_type {
